[refactor] Undo, Redo도 외부로 분리, 디렉토리 노드도 분리
