+++ Engine/src/Editor/Scripting/ScriptReloadHelpers.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace Alice
+{
+	class World;
+
+	bool ReloadScripts_FromButton(World& world);
+}
